fix: bitflyerからの数値を受け取る際、少数な場合があったので修正

+++ api/models/models.go
@@ -16,15 +16,15 @@ type Ticket struct {
 	State           string  `json:"state"`
 	Timestamp       string  `json:"timestamp"`
 	TickID          int     `json:"tick_id"`
-	BestBid         int     `json:"best_bid"`
-	BestAsk         int     `json:"best_ask"`
+	BestBid         float64 `json:"best_bid"`
+	BestAsk         float64 `json:"best_ask"`
 	BestBidSize     float64 `json:"best_bid_size"`
-	BestAskSize     int     `json:"best_ask_size"`
+	BestAskSize     float64 `json:"best_ask_size"`
 	TotalBidDepth   float64 `json:"total_bid_depth"`
-	TotalAskDepth   int     `json:"total_ask_depth"`
-	MarketBidSize   int     `json:"market_bid_size"`
-	MarketAskSize   int     `json:"market_ask_size"`
-	Ltp             int     `json:"ltp"`
+	TotalAskDepth   float64 `json:"total_ask_depth"`
+	MarketBidSize   float64 `json:"market_bid_size"`
+	MarketAskSize   float64 `json:"market_ask_size"`
+	Ltp             float64 `json:"ltp"`
 	Volume          float64 `json:"volume"`
 	VolumeByProduct float64 `json:"volume_by_product"`
 }
@@ -32,7 +32,7 @@ type Ticket struct {
 type Execution struct {
 	ID                         int     `json:"id"`
 	Side                       string  `json:"side"`
-	Price                      int     `json:"price"`
+	Price                      float64 `json:"price"`
 	Size                       float64 `json:"size"`
 	ExecDate                   string  `json:"exec_date"`
 	BuyChildOrderAcceptanceID  string  `json:"buy_child_order_acceptance_id"`
